requirements.txt
pyaudio
soundfile
SpeechRecognition[whisper-local]
coqui-tts
torch
python-dotenv